move grpc code and decouple from electron events
